Added exception handling to RequestedResourceServiceImpl. Fails RequestDescriptionTest.requestArtifactDescription.

+++ src/main/java/de/fraunhofer/isst/dataspaceconnector/services/resource/RequestedResourceServiceImpl.java
@@ -1,10 +1,14 @@
 package de.fraunhofer.isst.dataspaceconnector.services.resource;
 
 import de.fraunhofer.iais.eis.Resource;
+import de.fraunhofer.isst.dataspaceconnector.exceptions.ResourceException;
+import de.fraunhofer.isst.dataspaceconnector.exceptions.ResourceNotFoundException;
+import de.fraunhofer.isst.dataspaceconnector.exceptions.ResourceTypeException;
 import de.fraunhofer.isst.dataspaceconnector.model.RequestedResource;
 import de.fraunhofer.isst.dataspaceconnector.model.ResourceMetadata;
 import de.fraunhofer.isst.dataspaceconnector.services.IdsUtils;
 import de.fraunhofer.isst.dataspaceconnector.services.usagecontrol.PolicyHandler;
+import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,25 +28,37 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
     /** Constant <code>LOGGER</code> */
     public static final Logger LOGGER = LoggerFactory.getLogger(RequestedResourceServiceImpl.class);
 
-    private RequestedResourceRepository requestedResourceRepository;
-    private IdsUtils idsUtils;
+    private final RequestedResourceRepository requestedResourceRepository;
+    private final IdsUtils idsUtils;
 
-    private Map<UUID, Resource> requestedResources;
-    private PolicyHandler policyHandler;
+    private final Map<UUID, Resource> requestedResources;
+    private final PolicyHandler policyHandler;
 
-    @Autowired
     /**
      * <p>Constructor for RequestedResourceServiceImpl.</p>
      *
-     * @param requestedResourceRepository a {@link de.fraunhofer.isst.dataspaceconnector.services.resource.RequestedResourceRepository} object.
+     * @param requestedResourceRepository a {@link RequestedResourceRepository} object.
      * @param idsUtils a {@link de.fraunhofer.isst.dataspaceconnector.services.IdsUtils} object.
-     * @param policyHandler a {@link de.fraunhofer.isst.dataspaceconnector.services.usagecontrol.PolicyHandler} object.
+     * @param policyHandler a {@link PolicyHandler} object.
      */
-    public RequestedResourceServiceImpl(RequestedResourceRepository requestedResourceRepository, IdsUtils idsUtils, PolicyHandler policyHandler) {
+    @Autowired
+    public RequestedResourceServiceImpl(@NotNull RequestedResourceRepository requestedResourceRepository,
+                                        @NotNull IdsUtils idsUtils,
+                                        @NotNull PolicyHandler policyHandler) throws IllegalArgumentException{
+        if(requestedResourceRepository == null)
+            throw new IllegalArgumentException("The RequestedResourceRepository cannot be null.");
+
+        if(idsUtils == null)
+            throw new IllegalArgumentException("The IdsUtils cannot be null.");
+
+        if(policyHandler == null)
+            throw new IllegalArgumentException("The PolicyHandler cannot be null.");
+
         this.requestedResourceRepository = requestedResourceRepository;
         this.idsUtils = idsUtils;
         this.policyHandler = policyHandler;
 
+        // NOTE: This wont scale
         requestedResources = new HashMap<>();
         for (RequestedResource resource : requestedResourceRepository.findAll()) {
             requestedResources.put(resource.getUuid(), idsUtils.getAsResource(resource));
@@ -55,11 +71,10 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
      * Saves the resources with its metadata as external resource or internal resource.
      */
     @Override
-    public UUID addResource(ResourceMetadata resourceMetadata) {
-        RequestedResource resource = new RequestedResource(new Date(), new Date(), resourceMetadata, "", 0);
+    public UUID addResource(ResourceMetadata resourceMetadata) throws ResourceException {
+        final var resource = new RequestedResource(new Date(), new Date(), resourceMetadata, "", 0);
 
-        requestedResourceRepository.save(resource);
-        requestedResources.put(resource.getUuid(), idsUtils.getAsResource(resource));
+        storeResource(resource);
 
         return resource.getUuid();
     }
@@ -70,13 +85,14 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
      * Publishes the resource data.
      */
     @Override
-    public void addData(UUID resourceId, String data) {
-        RequestedResource resource = requestedResourceRepository.getOne(resourceId);
+    public void addData(UUID resourceId, String data) throws ResourceException{
+        final var resource = getResource(resourceId);
+        if(resource == null)
+            throw new ResourceNotFoundException("The resource does not exist.");
 
         resource.setData(data);
-        resource.setModified(new Date());
 
-        requestedResourceRepository.save(resource);
+        storeResource(resource);
     }
 
     /**
@@ -86,8 +102,12 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
      */
     @Override
     public void deleteResource(UUID resourceId) {
-        requestedResourceRepository.deleteById(resourceId);
-        requestedResources.remove(resourceId);
+        final var key = requestedResources.remove(resourceId);
+        if (key != null) {
+            requestedResourceRepository.deleteById(resourceId);
+        } else {
+            LOGGER.warn("Tried to delete resource that does not exist.");
+        }
     }
 
     /**
@@ -96,8 +116,18 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
      * Gets a resource by id.
      */
     @Override
-    public RequestedResource getResource(UUID resourceId) {
-        return requestedResourceRepository.getOne(resourceId);
+    public RequestedResource getResource(UUID resourceId) throws ResourceTypeException{
+        final var resource = requestedResourceRepository.findById(resourceId);
+
+        if (resource.isEmpty()) {
+            return null;
+        } else {
+            final var result = isValidResource(resource.get());
+            if (result.isPresent())
+                throw new ResourceTypeException("The resource is not valid. " + result.get());
+
+            return resource.get();
+        }
     }
 
     /**
@@ -106,8 +136,12 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
      * Gets resource metadata by id.
      */
     @Override
-    public ResourceMetadata getMetadata(UUID resourceId) {
-        return requestedResourceRepository.getOne(resourceId).getResourceMetadata();
+    public ResourceMetadata getMetadata(UUID resourceId) throws ResourceNotFoundException {
+        final var resource = getResource(resourceId);
+        if (resource == null)
+            throw new ResourceNotFoundException("This resource does not exist.");
+
+        return resource.getResourceMetadata();
     }
 
     /**
@@ -116,15 +150,20 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
      * Gets resource data by id.
      */
     @Override
-    public String getData(UUID resourceId) throws IOException {
-        RequestedResource resource = requestedResourceRepository.getOne(resourceId);
-        int counter = resource.getAccessed();
-
-        resource.setAccessed(counter + 1);
-        requestedResourceRepository.save(resource);
+    public String getData(UUID resourceId) throws ResourceException, IOException {
+        final var resource = getResource(resourceId);
+        if (resource == null)
+            throw new ResourceNotFoundException("This resource does not exist.");
 
         if (policyHandler.onDataAccess(resource)) {
-            return requestedResourceRepository.getOne(resourceId).getData();
+            final var data = resource.getData();
+            try {
+                storeResource(resource);
+            }catch(ResourceException exception){
+                throw new ResourceException("Failed to make the data access persistent.",
+                        exception);
+            }
+            return data;
         } else {
             return "Policy Restriction!";
         }
@@ -134,5 +173,30 @@ public class RequestedResourceServiceImpl implements RequestedResourceService {
     @Override
     public ArrayList<Resource> getRequestedResources() {
         return new ArrayList<>(requestedResources.values());
+    }
+
+    public Optional<String> isValidResource(RequestedResource resource) {
+        if (resource == null)
+            return Optional.of("The resource cannot be null.");
+
+        if (resource.getResourceMetadata() == null)
+            return Optional.of("The resource metadata cannot be null.");
+
+        if (resource.getResourceMetadata().getRepresentations() == null)
+            return Optional.of("The resource representation cannot be null.");
+
+        if (resource.getResourceMetadata().getRepresentations().size() < 1)
+            return Optional.of("The resource representation must have at least one element.");
+
+        return Optional.empty();
+    }
+
+    private void storeResource(RequestedResource resource) throws ResourceException {
+        final var result = isValidResource(resource);
+        if(result.isPresent())
+            throw new ResourceException("Not a valid resource. " + result.get());
+
+        requestedResourceRepository.save(resource);
+        requestedResources.put(resource.getUuid(), idsUtils.getAsResource(resource));
     }
 }
